Removed templates from delivery
